refactor(reports): extract pure helpers to reports_helpers.go
